+++ src/main/java/leets/leenk/domain/birthday/application/usecase/BirthdayUsecase.java
@@ -1,8 +1,8 @@
-package leets.leenk.domain.birthday.usecase;
+package leets.leenk.domain.birthday.application.usecase;
 
-import leets.leenk.domain.birthday.dto.response.BirthdayUserResponse;
-import leets.leenk.domain.birthday.mapper.BirthdayMapper;
-import leets.leenk.domain.birthday.service.BirthdayGetService;
+import leets.leenk.domain.birthday.application.dto.response.BirthdayUserResponse;
+import leets.leenk.domain.birthday.application.mapper.BirthdayMapper;
+import leets.leenk.domain.birthday.domain.service.BirthdayGetService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
