+++ src/main/java/leets/leenk/domain/birthday/presentation/BirthdayController.java
@@ -1,9 +1,9 @@
-package leets.leenk.domain.birthday.controller;
+package leets.leenk.domain.birthday.presentation;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import leets.leenk.domain.birthday.dto.response.BirthdayUserResponse;
-import leets.leenk.domain.birthday.usecase.BirthdayUsecase;
+import leets.leenk.domain.birthday.application.dto.response.BirthdayUserResponse;
+import leets.leenk.domain.birthday.application.usecase.BirthdayUsecase;
 import leets.leenk.global.common.response.CommonResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -12,7 +12,8 @@ import org.springframework.web.bind.annotation.RestController;
 
 import java.util.List;
 
-import static leets.leenk.domain.user.presentation.ResponseCode.GET_BIRTHDAY_USERS;
+import static leets.leenk.domain.birthday.presentation.ResponseCode.GET_BIRTHDAY_USERS;
+
 
 @Tag(name = "BIRTHDAY")
 @RestController
